requirements.txt
flask
pymongo
flask-jwt-extended
flask-cors
python-dotenv
werkzeug
flask-mail
google-auth